+++ 01-Code/LIFEv2/LayerContainer/LayerContainerFacade/Implementation/LayerContainerFacadeImpl.cs
@@ -61,7 +61,7 @@ namespace LayerContainerFacade.Implementation {
         public bool InitializeLayer(TLayerInstanceId instanceId, TInitData initData) {
             return _rteManager.InitializeLayer(instanceId, initData);
         }
-        public long Tick(int amountOfTicks = 1)
+        public long Tick(long amountOfTicks = 1)
         {
             return _rteManager.Advance(amountOfTicks);
         }
